Lets implement the sermon notes section as this is the next crucial piece. We need this to be a bit more robust. As church members are listening to the sermon live, they should be able to take notes. They also should be able to retrieve, update and delete notes from previous sermons. Search the sermon topic, or the speaker of the sermon.
We also will use this sermon when using our chat. The chat will be able to search sermons for relevant content and allow users to be able to discuss and breakdown the sermons as well.

We also will have the capability to listen to a sermon from youtube and take notes from that sermon all within the same window or sidebar