The notes on the sermon details page is fine. However this assumes that we have a sermon already. We need the ability to add a new sermon from the resources page so that we can have a details page.
We also should be able to have a dedicated sermon chat directly about the sermon within the details page. The chat will take context from the sermon notes, scripture, title. It can auto generate a summary after the user enters in a few notes and the scripture verse along with the title. if all 3 of these are in place then the summary is auto generated.

 We need an intuitive way to integrate a chat on the details so it is clean and minimalistic.. Create and design this so you will have to come up with a good idea as most chats take up the entire screen and we dont have that luxury on this details screen. So we will have a dedicated sermon chat and then of course our general chat that encompasses everything.