Lets go ahead an implement pinecone logic as well as add our GOOGLE_API_KEY to our genkit. We will use gemini flash as our model
The pinecone api keys are in our google secrets api

Under
PINECONE_API_KEY
and for the index that uses pinecone vector we use
PINECONE_INDEX_URL

These should be good to use with our chat in both the chat screen as well as the sermon details

Lets make sure this is production ready. No stubs no psuedo code or mock data on the chat screens. The chat at least should be production ready.